projeto_datathon_brunag/src/requirements.txt
anyio==4.2.0
beautifulsoup4==4.12.3
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
evidently==0.2.2
fastapi==0.115.13
filetype==1.2.0
gdown==5.2.0
httpcore==1.0.9
httpx==0.27.0
idna==3.7
importlib-metadata==7.1.0
imbalanced-learn==0.12.0
joblib==1.4.2
numpy==1.26.4
openpyxl==3.1.5
packaging==24.0
pandas==2.2.2
pdfplumber==0.10.4
Pillow==10.3.0
protobuf==5.27.0
py7zr==1.0.0
pyarrow==12.0.1
pycryptodomex==3.20.0
Pygments==2.18.0
python-json-logger>=2.0.4
python-multipart==0.0.20
pytz==2025.2
pyzstd==0.17.0
requests==2.32.4
scikit-learn==1.3.2
shap<0.50.0,>=0.40.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
texttable==1.7.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
uvicorn==0.34.3
watchdog==6.0.0
xgboost>=1.7.6
PyYAML==5.1.2
pytest>=7.0
unidecode
